Integrated with external api to fetch job listings from the API

+++ app/Http/Controllers/ListingController.php
@@ -3,20 +3,30 @@
 namespace App\Http\Controllers;
 
 use App\Models\JobListings;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Http;
 use Illuminate\Validation\Rule;
 
 
 class ListingController extends Controller
 {
 
+
+
+
+
+
+
     //retrieve all job listings from db
     public function index(){
-
         return view('listings.index', [
             'listings'=> JobListings::latest()->
-            filter(request(['tag', 'search']))->paginate(6)
+            filter(request(['tag', 'search']))->paginate(10)
         ]);
+
+
+
     }
 
     //get Single JobListing from db
@@ -51,6 +61,8 @@ class ListingController extends Controller
             $validatedData['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $validatedData['user_id'] = auth()->id();
+
         JobListings::create($validatedData);
 
         return redirect('/')->with('message', 'Job Listed Successfully!');
@@ -65,6 +77,13 @@ class ListingController extends Controller
 
     //update job listing
     public function update(Request $request, JobListings $listing){
+
+        //make sure logged user is owner of job listing
+
+        if($listing->user_id !== auth()->id()){
+            abort(403, 'Unauthorized action.');
+        }
+
 
         $validatedData = $request->validate([
             'title' => 'required',
@@ -85,13 +104,26 @@ class ListingController extends Controller
 
         $listing->update($validatedData);
 
-        return back()->with('message', 'Updated Successfully!');
+        return redirect('/')->with('message', 'Job Updated Successfully!');
     }
 
     //delete job listing
     public function destroy(JobListings $listing){
+        //make sure logged user is owner of job listing
+
+        if($listing->user_id !== auth()->id()){
+            abort(403, 'Unauthorized action.');
+        }
         $listing->delete();
         return redirect ('/')->with('message', 'Deleted Successfully!');
+    }
+
+
+    //show all job listings by user
+    public function manage(){
+        return view('listings.manage', [
+            'listings'=>auth()->user()->listings()->get()
+        ]);
     }
